// apps/web/src/lib/api/generated/model/workScheduleSlotResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, billing, products
 * OpenAPI spec version: 1.0.0
 */
import type { WorkScheduleSlotResponseWorkday } from './workScheduleSlotResponseWorkday';
import type { WorkScheduleSlotResponseWorkHour } from './workScheduleSlotResponseWorkHour';

export interface WorkScheduleSlotResponse {
  workday: WorkScheduleSlotResponseWorkday;
  workHour: WorkScheduleSlotResponseWorkHour;
}